+++ common/socks5-client-core/src/error.rs
@@ -29,7 +29,7 @@ impl From<ConnectionError> for Socks5ClientCoreError {
     }
 }
 
-impl nym_task::manager::TaskStatusEvent for Socks5ClientCoreError {
+impl nym_task::TaskStatusEvent for Socks5ClientCoreError {
     fn as_any(&self) -> &dyn std::any::Any {
         self
     }
