+++ common/socks5-client-core/src/lib.rs
@@ -23,8 +23,7 @@ use nym_client_core::init::types::GatewaySetup;
 use nym_credential_storage::storage::Storage as CredentialStorage;
 use nym_sphinx::addressing::clients::Recipient;
 use nym_sphinx::params::PacketType;
-use nym_task::manager::TaskStatus;
-use nym_task::{TaskClient, TaskHandle};
+use nym_task::{TaskClient, TaskHandle, TaskStatus};
 
 use anyhow::anyhow;
 use nym_validator_client::UserAgent;
